app-requirements.txt
streamlit>=1.33
pandas>=2.0
numpy>=1.24
duckdb>=0.9
matplotlib>=3.7
python-dotenv>=1.0
langchain>=0.2
langchain-community>=0.2
langchain-google-genai>=1.0
langchain-experimental>=0.2
scikit-learn>=1.3
statsmodels>=0.14
tabulate>=0.9
databricks-sql-connector>=2.14
